+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - nodejs: restore from history
 - python: restore from history
 - ruby: restore from history
+- php: restore from history
 
 ## 1.6.0
 - nodejs: setUsers for region protection